requirements.txt
chromadb<0.4,>=0.3
deeplake<4,>=3.6
google-api-python-client<3,>=2.88
google-auth-httplib2<1,>=0.1
google-auth-oauthlib<2,>=1.0
google-cloud-aiplatform<2.0,>=1.25.0
grpcio<2,>=1.54
langchain[llms]<0.1,>=0.0.190
pillow<10,>=9.5
pypdf2<4,>=3.0
sentence-transformers<3,>=2.2
transformers<5,>=4.29
